requirements.txt
# requirements.txt
streamlit
transformers
tweepy
python-dotenv
requests
matplotlib
pandas
